Move /diff to its own router
